requirements.txt
Flask
tensorflow
keras
Pillow
gunicorn
